Added Coordinator, Modified Simulation Result
After the failure with the Batch Simulator, I had another idea. 
+ Added Cooridnatior
[n] Coordinator coordinates and gathers results from multiple Simulator instances, running a sub-set of games.
+ I extracted and modified the Simulation Result. I added Result Merge logic, and allowed bypassing of the initial calculations.

THe Coordinator beats the BatchSimuilator by 10x and the Simulator alone by ~6x. I am pleased. 
I am considering dropping the number of processes specifier, in favor of letting the language determine the best amount. Be warned, the Coordinator is resource hungry.

I also need to find a better way to configure the simulations etc. And to log progress better. Lastly I need to ensure that at least the given number of games are run.